Some Audio, New Hand
added 2 cursed music tracks, 6 variations of the same sound and a thumbs up. Grab SFX could just be randomly selected (if it's easy just pick one if easier) when you grab the torch. I can do a separate one for whiffed grab if it feels like it needs it. 
Also: Noodlier arm