Audio + PUNCH BUG
Fire is crackling and music is playing!
Added a line into the Grabbable script to play Grab01 sound when grab is successful.

There's a bug where if you don't grab the torch, the collision doesn't leave your hand, meaning you can only slap it around and never grab it again.